refactor: move functions to helpers in signIn
