+++ src/app/auth/controllers/singInUser.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/gofiber/fiber/v2"
 	authDtos "github.com/rodrigoRVSN/beeus-api/src/app/auth/dtos"
-	authService "github.com/rodrigoRVSN/beeus-api/src/app/auth/services"
+	authService "github.com/rodrigoRVSN/beeus-api/src/app/auth/services/signInUser"
 )
 
 func injectTokenCookie(context *fiber.Ctx, user *authDtos.SignInOutput) {
