+++ src/app/auth/services/signInUser/helpers/findUserByEmail.go
@@ -1,0 +1,16 @@
+package authService
+
+import (
+	"github.com/rodrigoRVSN/beeus-api/src/app/users/entities"
+	database "github.com/rodrigoRVSN/beeus-api/src/infra/config"
+)
+
+func FindUserByEmail(email string) (*entities.User, error) {
+	var user entities.User
+
+	if err := database.DB.First(&user, "email = ?", email).Error; err != nil {
+		return nil, err
+	}
+
+	return &user, nil
+}
